chore: reorganized client-pages component folder
