+++ components/client-pages/search/SearchReportsClient.tsx
